+++ src/interceptor_pre/create_error_response.py
@@ -1,4 +1,4 @@
-from src.constants import API_CALL_ERROR, JWT_ERROR, NO_EVENT_ERROR, NO_DATA_ERROR, BIOT_SERVICE_ENVIRONMENT, BIOT_APP_NAME
+from src.constants import API_CALL_ERROR, JWT_ERROR, NO_EVENT_ERROR, NO_DATA_ERROR, ACCESS_DENIED_ERROR, BIOT_SERVICE_ENVIRONMENT, BIOT_APP_NAME
 from src.utils.configure_logger import logger
 from src.utils.traceparent_utils import parse_traceparent_string
 
@@ -49,6 +49,15 @@ def no_data_error(error, trace_id):
         trace_id=trace_id
     )
 
+def access_denied_error(error, trace_id):
+    return build_response_error(
+        status_code=403,
+        code=ACCESS_DENIED_ERROR,
+        message="Access is denied",
+        trace_id=trace_id
+    )
+
+
 def internal_server_error(error, trace_id):
     return build_response_error(
         status_code=500,
@@ -57,19 +66,22 @@ def internal_server_error(error, trace_id):
         trace_id=trace_id
     )
 
+
 errors = {
     API_CALL_ERROR: api_error,
     JWT_ERROR: jwt_error,
     NO_EVENT_ERROR: no_event_error,
     NO_DATA_ERROR: no_data_error,
+    ACCESS_DENIED_ERROR: access_denied_error,
     "internalServerError": internal_server_error
 }
+
 
 def create_error_response (error, traceparent):
     logger.error("Got error: ", error)
     trace_id = parse_traceparent_string(traceparent)
-    if error is not None and error in errors:
-        return errors[error](error, trace_id)
+    if error is not None and error.args is not None and error.args[0] in errors.keys():
+        return errors[error.args[0]](error, trace_id)
     else:
         return errors["internalServerError"](error, trace_id)