+++ src/nonspecific/extract_data_from_event.py
@@ -1,16 +1,16 @@
 import json
 from src.constants import NO_EVENT_ERROR, NO_DATA_ERROR, JWT_ERROR
 
-def extract_data_from_event (event): 
+
+def extract_data_from_event(event):
     if event is None:
         raise Exception(NO_EVENT_ERROR)
 
-    parsed_event = json.loads(event["body"])
+    body_str: str = event["body"]
+    parsed_event = json.loads(body_str.replace("\\n", ""))
 
     if parsed_event is None:
         raise Exception(NO_DATA_ERROR)
 
-    return {
-        "data": parsed_event["data"] if "data" in parsed_event else None
-    }
+    return parsed_event